Turn on preemption for tracepoints probes
In current versions of the kernel (up to 5.18) the functions attached to
tracepoints are called with preemption disabled (see definition of
__DO_TRACE macro in include/linux/tracepoint.h). This effectively means
that probe functions cannot sleep and thus must not use mutexes, access
userspace data etc. On kernels 5.16+ (due to some default config changes)
attempts to do so will cause random "BUG: scheduling while atomic"
errors.
This commit turns preemption on for our probe functions based on the
assumption that all the tracepoints we attach to are called from the
user context, and because of that it should be safe to be preempted
there.
Preemption is still turned off, as required, for smp_processor_id().

This patch will probably not be needed in some future after sleepable
tracepoints are introduced to the mainline kernel.