Rust sync (#157)
* rust-ftdb: API cleanup: iterators

This patch:
- Adds xxx_iter() methods when xxx() returns Vec
- Updates methods to return "impl ExactSizeIterator" if size is known
- Adds missing size_hint() methods for custom iterators
- Fixes bugs in existing size_hint() implementations
- Adds implementation of ExactSizeIterator to custom iterators
- Fixes offset-by-one issue in Iterators
- Fixes issues reported by clippy

Version bump: 0.5.0

* rust-ftdb: Added db-hash example that matches C FTDB example

* rust-ftdb: Add switches to FunctionEntry

Version bump: 0.5.1

* rust-ftdb: Add Location wrapper over raw &str

Version bump: 0.6.0

* rust-ftdb: Add plain_id and try_plain_id methods to CallEntryRefInfo

Version bump: 0.6.1

* rust-ftdb: Add DFS iterator for Functions

Version bump: 0.6.2

* rust-ftdb: Add query interfaces to extend FFI types with methods

ftdb-sys version bump: 0.7.0

* rust-ftdb: Introduce identifiers with db serialization

This patch:
- Adds FunctionId, GlobalId, TypeId newtypes wrapping raw integers
- Adds serialization to uint pg types (requires pguint db extension)
  that is enabled by a feature 'postgres', enabled by default.

Version bump: 0.6.3

* rust-ftdb: Improve API design for async use cases

This patch reorganizes internal API by introducing borrowed and owned
types Borrowed type is the new name for previous ftdb types
implementation. It's a simple pointer to FFI structure with <'a>
lifetime that depends on Ftdb instance.

Owned type might exists without the need of Ftdb instance. With this
type internal reorganization has been done. A new type is introduced:
FtdbHandle which is actual owner of a FFI pointer. As long as the
FtdbHandle instance exists, FFI memory won't be freed. With FtdbHandle
comes another helper type called Owned<T>. This type stores a pointer
to FFI structure and atomic shared reference (Arc) to FtdbHandle.
This means that cloning Owned instance (therefore increasing counter
of references to FtdbHandle instance) is a safe way to ensure that
FFI memory won't be released. When all references to FtdbHandle are
dropped, the memory is released.

This is a big change from previous implementation as now `Ftdb` is not
a real owner of FFI pointers but one of its users. It is possible to
get inner structures (Fops, Functions, some IntoIterators) and drop
actual Ftdb instance and FFI memory won't be released.

This change has been required to use `ftdb` API with tokio in async
environment. When using `tokio::spawn()` it is required that closure
passed to this function is Sync + Send + 'static. This means that all
captured references cannot be of lifetimes other than 'static. When
using previous API, for example when iterating over `FunctionEntry<'a>`
objects and processing them in separate threads with `tokio::spawn`
it was not possible to compile such program because <'a> lifetime means
that FunctionEntry is of lifetime shorter than 'static. With new change
there is another type OwnedFunctionEntry that doesn't have any external
lifetimes because it holds an `Arc<FtdbHandle>` itself.

Keep in mind that the API polishing is not finished yet as it's being
updated while working on project using it (and discovering problems
that way).

This patch also adds several helper methods to make the use of it a bit
easier. For example instead of calling `db.funcs().iter()` you can call
`db.funcs_iter()` directly.

Version bump: 0.7.0

---------

Co-authored-by: Blazej Jaworowski <b.jaworowski@samsung.com>