parser: ignore pid 0 tracer entries
PID 0 is effectively used nowhere important in regards to what user sees,
it is a placeholder for kernel operations in userspace, a hand of god, so
to speak.

In tracer, entries with PID 0 are outputted as Env entries, the PID is
just a placeholder to not break the format. This however, was not taken
into account by parser which logged PID 0 events and caused breakage for
processes that use root pid.

This patch is backported from tracer/env branch