Change directory structure and split CMake to different files (#259)
* ftdb: restructure directories and CMake

FTDB is conceptually a different project and even though the code is
similar, having everything in BAS directory made life a bit difficult.

We introduce some duplication (in particular ftdb_entry.h and maps.c/h),
but FTDB has a possible refactoring planned so this may change in the
future.

Additionaly, we added CMake and reduced LoC in root CMakeLists.txt.

Logically, It would also make sense to put rust FTDB bindings here, but
again thats for future.

* clang-proc: simplify and split cmake to clang-proc directory

Most of the code in root CMakeLists.txt was related to settings up LLVM
for clang-proc.

LLVM has changed a bit for the better with regards to CMake.
Now find_package(LLVM) finds llvm-config for us, it also sets up some
useful variables. This means that we don't need to define LLVM_CONFIG_BIN
when building!

The CMake LLVM API has a function for translating components to needed
libraries, but that does not work for us (I think so?), so
cmake/llvm-config.cmake is supposed to do a full setup of all the
variables we need to compile clang-proc

* bas: split compile targets to different directories

bas/ directory previously was a home to many different things, which may
have been a bit confusing.

Each compile target has its own directory with its own CMakeLists.txt
file.

* cmake: restore ability to specify LLVM_CONFIG_BIN

There are some cases in which you want to use a specific version of LLVM.
For such cases you can again specify LLVM_CONFIG_BIN yourself.

If LLVM_CONFIG_BIN is not specified, then find_package() will try to find
LLVM by itself.

* clang-proc: set -O2 in release build only in clang-proc

* clang-proc/cmake: correct nonsense code

* cmake: update llvm version according to specified llvm-config

LLVM_PACKAGE_VERSION is a variable set by find_package(LLVM).
We use it regardless of whether we use LLVM provided manually or if its
found by CMake. This caused a mismatch between reported LLVM version and
the used one.

* cmake: default to llvm-config provided in path