pyftdb: code cleanup (#281)
* pyftdb: unexport internal types

PyFTDB has various internal types exported whose constructors are
expecting valid data. They were not really meant to be exported, so we
are reverting this commit.

Along this change, lets make type registering look a bit less spammy.

* pyftdb: remove dead code

Some of this code was still referring to data used in nfsdb and was not
used in ftdb at all. Let's not duplicate it and remove it.

* pyftdb: reduce allocation size redundancy and remove repeated types

This one is quite funny. .tp_basicsize field in PyTypeObject defines the
size of 'private data' for each method implemented in natively in
CPython for a certain object. It also defines how much bigger the
allocation should be to account for this object specific data.

Here we almost everywhere use the sizeof(PyTypeObject) itself, which is
419 at the time of writing for Python 3.12.

We only use 16 bytes (out of which 8 bytes is redundant) because the
actual data types used are way smaller.

Also, there are quite a lot of redefinitions for a type with the same
memory layout for each object specific data so they have been removed in
place of libftdb_ftdb_collection_object.

* fix: removed typings absent in libftdb

* Simplified some duplicated code for collection types and iterators.

* Added custom type for unresolvedfuncs entry.

---------

Co-authored-by: l.topolski <l.topolski@samsung.com>
Co-authored-by: Mateusz Manko <m.manko@samsung.com>