tracer: use per-cpu prefix tries (#282)
* tracer: use per-cpu prefix tries

It turns out, unsuprisingly so, that using a big lock on a pretty
contented resource causes unnecessary overhead.

This problem revealed itself in much bigger builds. I tested env tracing
on a kernel build (with 16 threads) and it behaved fine.

The problem came when we tried to compile AOSP platform, which uses not
only C, but obviously Java, C++, Kotlin, Rust and whoknows what more.

This probably resulted in bigger env variables, causing longer inserts on
a bigger amount of CPU's (AOSP builds were done on 128 core machines)
which in turn caused contentation.

This commit decentralizes env tracing to each cpu instead of having one
big lock on a one big tree.

* tracer: enable interrupts and preemption in fail path

* tracer: don't disable interrupts on per cpu access

The decision to disable interrupts was a bit of a mistake on my part.

It does not do really anything in regards to how the tracer will
behave during per cpu trie access when the interrupt hits, because due to
disabled preemption it couldn't have been moved to a different CPU either
way.

This on the other hand, introduced another bug, thankfully found by
looking at the code at not actually triggering it.

trie.h hardcodes gfp_t flags to GFP_KERNEL, which is improper in context
of disabled preemption and interrupts. If kmalloc() sleeps, then this
could've caused a CPU core deadlock.

* tracer: change gfp flags to GFP_NOWAIT in atomic context